auth/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.2
django-model-utils==4.4.0
django-oauth-toolkit==2.3.0
idna==3.6
jwcrypto==1.5.4
oauthlib==3.2.2
pycparser==2.21
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
